Created services, and added flashcards loading/saving service
